# Conflicts:
#	logic/src/commonMain/kotlin/com/wire/kalium/logic/feature/UserSessionScope.kt
#	logic/src/commonMain/kotlin/com/wire/kalium/logic/sync/receiver/handler/legalhold/LegalHoldHandler.kt
#	logic/src/commonTest/kotlin/com/wire/kalium/logic/sync/receiver/handler/legalhold/LegalHoldHandlerTest.kt

+++ logic/src/androidMain/kotlin/com/wire/kalium/logic/feature/message/GetPaginatedFlowOfMessagesByConversationUseCase.kt
@@ -40,7 +40,7 @@ class GetPaginatedFlowOfMessagesByConversationUseCase internal constructor(
     suspend operator fun invoke(
         conversationId: ConversationId,
         visibility: List<Message.Visibility> = Message.Visibility.values().toList(),
-        startingOffset: Int,
+        startingOffset: Long,
         pagingConfig: PagingConfig
     ): Flow<PagingData<Message.Standalone>> = messageRepository.extensions.getPaginatedMessagesByConversationIdAndVisibility(
         conversationId, visibility, pagingConfig, startingOffset
